+++ apps/parent/src/main/java/com/instructure/parentapp/di/feature/AssignmentDetailsModule.kt
@@ -25,7 +25,6 @@ import com.instructure.pandautils.features.assignments.details.AssignmentDetails
 import com.instructure.pandautils.features.assignments.details.AssignmentDetailsRepository
 import com.instructure.pandautils.features.assignments.details.AssignmentDetailsRouter
 import com.instructure.pandautils.features.assignments.details.AssignmentDetailsSubmissionHandler
-import com.instructure.pandautils.receivers.alarm.AlarmReceiverNotificationHandler
 import com.instructure.pandautils.room.appdatabase.daos.ReminderDao
 import com.instructure.pandautils.utils.ColorKeeper
 import com.instructure.parentapp.features.assignment.details.ParentAssignmentDetailsBehaviour
@@ -33,7 +32,6 @@ import com.instructure.parentapp.features.assignment.details.ParentAssignmentDet
 import com.instructure.parentapp.features.assignment.details.ParentAssignmentDetailsRepository
 import com.instructure.parentapp.features.assignment.details.ParentAssignmentDetailsRouter
 import com.instructure.parentapp.features.assignment.details.ParentAssignmentDetailsSubmissionHandler
-import com.instructure.parentapp.features.assignment.details.receiver.ParentAlarmReceiverNotificationHandler
 import com.instructure.parentapp.util.ParentPrefs
 import com.instructure.parentapp.util.navigation.Navigation
 import dagger.Module
@@ -41,7 +39,6 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.FragmentComponent
 import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(FragmentComponent::class)
@@ -79,14 +76,5 @@ class AssignmentDetailsModule {
     @Provides
     fun provideAssignmentDetailsColorProvider(parentPrefs: ParentPrefs, colorKeeper: ColorKeeper): AssignmentDetailsColorProvider {
         return ParentAssignmentDetailsColorProvider(parentPrefs, colorKeeper)
-    }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-class AssignmentDetailsSingletonModule {
-    @Provides
-    fun provideAssignmentDetailsNotificationHandler(): AlarmReceiverNotificationHandler {
-        return ParentAlarmReceiverNotificationHandler()
     }
 }
