+++ apps/teacher/src/main/java/com/instructure/teacher/di/AlarmSchedulerModule.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2024 - present Instructure, Inc.
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, version 3 of the License.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+package com.instructure.teacher.di
+
+import com.instructure.pandautils.receivers.alarm.AlarmReceiverNotificationHandler
+import com.instructure.teacher.features.reminder.TeacherAlarmReceiverNotificationHandler
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+class AlarmSchedulerModule {
+    @Provides
+    fun provideAssignmentDetailsNotificationHandler(): AlarmReceiverNotificationHandler {
+        return TeacherAlarmReceiverNotificationHandler()
+    }
+}
