+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/calendarevent/details/EventViewModel.kt
@@ -85,6 +85,11 @@ class EventViewModel @Inject constructor(
         loadData()
     }
 
+    override fun onCleared() {
+        super.onCleared()
+        reminderManager.removeLiveDataObserver()
+    }
+
     private fun loadData() {
         setToolbarColor()
         viewModelScope.tryLaunch {
